pyproject.toml
[tool.poetry]
name = "green"
version = "0.1.0"
description = ""
authors = ["FiazAbrar <abrarfiaz.bd@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
psycopg2-binary = "^2.9.1"
python-decouple = "^3.4"
djangorestframework = "^3.12.4"
Pillow = "^8.3.1"
django-cors-headers = "^3.8.0"
dj-rest-auth = "^2.1.11"
django-allauth = "^0.45.0"
django-extensions = "^3.1.3"
django-versatileimagefield = "^2.2"
gunicorn = "^20.1.0"
django-storages = "^1.11.1"
boto3 = "^1.18.42"
django-ckeditor = "^6.1.0"
django-csp = "^3.7"
django-json-widget = "^1.1.1"
sentry-sdk = "^1.4.3"
pyparsing = "^2.4.7"
pydot = "^1.4.2"
django-debug-toolbar = "^3.2.2"
openpyxl = "^3.0.9"

[tool.poetry.dev-dependencies]
black = {version = "^21.12b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "mirrrrr"
url = "https://mirrors.sustech.edu.cn/pypi/simple"
default= true